# api/docs/swagger.yaml
definitions:
  models.AttachPermission:
    properties:
      permission_id:
        type: string
      role_id:
        type: string
    required:
    - permission_id
    - role_id
    type: object
...
  models.PermissionModel:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  models.RoleModel:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.PermissionModel'
        type: array
    type: object
...
  models_v1.Response:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:8123
info:
  contact: {}
  title: E-commerce project v2
paths:
...
  /api/roles/attach:
    delete:
      consumes:
      - application/json
      description: DisAttach Permission To Role
      operationId: DisAttachPermissionToRole
      parameters:
      - description: disattach body
        in: body
        name: disattach_body
        required: true
        schema:
          $ref: '#/definitions/models.AttachPermission'
      produces:
      - application/json
      responses:
        "200":
          description: Disattach Permission To Role
          schema:
            $ref: '#/definitions/models.AttachPermission'
        "400":
          description: Bad request / bad UUID
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: DisAttach Permission To Role
      tags:
      - role
    post:
      consumes:
      - application/json
      description: attach permission to role
      operationId: AttachPermissionToRole
      parameters:
      - description: attach body
        in: body
        name: attach_body
        required: true
        schema:
          $ref: '#/definitions/models.AttachPermission'
      produces:
      - application/json
      responses:
        "200":
          description: Attach Permission To Role
          schema:
            $ref: '#/definitions/models.AttachPermission'
        "400":
          description: Bad request / bad UUID
          schema:
            $ref: '#/definitions/models_v1.Response'
        "404":
          description: Role not found/ Permission not found
          schema:
            $ref: '#/definitions/models_v1.Response'
        "409":
          description: Already exists
          schema:
            $ref: '#/definitions/models_v1.Response'
        "500":
          description: Internal
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: attach permission to role
      tags:
      - role
  /api/roles/permission:
    get:
      consumes:
      - application/json
      description: get all permissions in db
      operationId: getAllPermissions
      produces:
      - application/json
      responses:
        "200":
          description: Permissions list
          schema:
            items:
              $ref: '#/definitions/models.PermissionModel'
            type: array
        "500":
          description: Internal
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: get all permissions
      tags:
      - role
  /api/roles/role:
    get:
      consumes:
      - application/json
      description: get all roles in db
      operationId: getAllRoles
      produces:
      - application/json
      responses:
        "200":
          description: Roles list
          schema:
            items:
              $ref: '#/definitions/models.RoleModel'
            type: array
        "500":
          description: Internal
          schema:
            $ref: '#/definitions/models_v1.Response'
      security:
      - ApiKeyAuth: []
      summary: get all roles
      tags:
      - role
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"